ValidationEvent middleware refactor
Events validations has been moved from routes and error handling reference to these validations has been moved as well to the middleware module
